requirements.dev.txt
coverage
pylama
pytest
pytest-aiohttp
pytest-cov
sqlalchemy_utils
